Update Prisma library and add postinstall hook
The Prisma Client depends on a query engine binary that is generated
whenever `prisma generate` is run. The query engine is responsible for
managing database connections, generating SQL queries, and processing
responses from the database. The binary should be updated whenever the
Prisma schema changes. This commit adds `prisma generate` to the
postinstall scripts which should be triggered with each new version
that is pushed to the SIBR server.